requirements.txt
# AI Voice Assistant - Minimal Dependencies
# Core voice processing
deepgram-sdk==4.1.0
pyaudio==0.2.14
websockets==12.0

# AI and chat
openai==1.51.0
groq==0.9.0
httpx==0.27.2

# Google Calendar integration
google-auth==2.38.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.161.0

# Date/time handling
pytz==2023.4
python-dateutil==2.9.0.post0

# Environment and configuration
python-dotenv==1.0.1

# Document generation (optional - for transcript saving)
python-docx==1.2.0

# GUI (built into Python, but listed for clarity)
# tkinter is built-in to Python

# Standard library dependencies (no installation needed):
# - datetime
# - threading  
# - asyncio
# - json
# - os
# - pathlib
# - re
# - uuid
# - webbrowser
# - http.server
# - urllib.parse
# - shutil
# - subprocess
# - sys
